requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.105.0
h11==0.14.0
idna==3.6
pydantic==2.5.2
pydantic_core==2.14.5
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
aiogram==3.1.1
python-dotenv==1.0.0
requests==2.31.0
fastapi-users[sqlalchemy]
SQLAlchemy==2.0.22
asyncpg==0.28.0
alembic==1.12.0
psycopg2-binary==2.9.9
redis==5.0.1